+++ pkg/key/key.go
@@ -2,6 +2,4 @@ package key
 
 const (
 	Finalizer              = "giantswarm.io/logging-operator"
-	LoggingLabel           = "giantswarm.io/logging"
-	NetworkMonitoringLabel = "giantswarm.io/network-monitoring"
 )